Clean up renaming of name_extractor
