+++ examples/anime_names.py
